Moved run_cmd into __init__.py
Removed `tools/cmds.py`. Adjusted importing of `run_cmd`.
